requirements.txt
requests==2.21.0
selenium==3.141.0
youtube_dl==2019.6.27
httplib2==0.12.3
google_auth_oauthlib==0.3.0
numpy==1.16.3
moviepy==1.0.0
Pillow==6.0.0
google_api_python_client==1.7.9
praw==6.3.1
